Fix bug when answering correct.
When answering correct, capital would get removed from the dict.
This caused weird behaviour if the same country was chosen again.
It could also cause infinite loops,
if no countries have capitals in their dict anymore.
This is solved by copying the list with capitols,
instead of manipulating it directly.
This is because lists/dicts are mutable.